+++ src/app/core/data/src/main/java/g/sig/core_data/database_impl/dao/MonthDao.kt
@@ -3,7 +3,6 @@ package g.sig.core_data.database_impl.dao
 import androidx.room.*
 import g.sig.core_data.models.transaction.Month
 import g.sig.core_data.models.transaction.MonthCategories
-import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface MonthDao {
@@ -12,12 +11,12 @@ interface MonthDao {
     suspend fun setMonth(month: Month)
 
     @Query("SELECT budget FROM Month WHERE monthId == :monthId")
-    suspend fun getBudget(monthId: Int): Double?
+    suspend fun getBudget(monthId: Long): Double?
 
     @Query("SELECT expenses FROM Month WHERE monthId == :monthId")
-    suspend fun getExpenses(monthId: Int): Double?
+    suspend fun getExpenses(monthId: Long): Double?
 
     @Transaction
     @Query("SELECT * FROM Month WHERE monthId == :monthId")
-    suspend fun getCategories(monthId: Int): MonthCategories?
+    suspend fun getCategories(monthId: Long): MonthCategories?
 }
